+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Compatibility:
 * Handle either positional or keywords arguments by default in `Struct.new` (#3039, @rwstauner).
 * Promote `Set` class to core library (#3039, @andrykonchin).
 * Support `connect_timeout` keyword argument to `TCPSocket.{new,open}` (#3421, @manefz, @patricklinpl, @nirvdrum, @rwstauner).
+* Add `File.lutime` and `Pathname#lutime` methods (#3039, @andrykonchin).
 
 Performance: